backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
tensorflow-cpu
keras
numpy
websockets
mediapipe
onnxruntime
huggingface_hub
